Refactor exception handling - delete

+++ src/main/java/com/treinandospring/course/services/UserService.java
@@ -33,14 +33,15 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		try {
-
-			repository.deleteById(id);
-
-		} catch (EmptyResultDataAccessException e) {
+		Optional<User> user = repository.findById(id);
+		if (user.isPresent()) {
+			try {
+				repository.deleteById(id);
+			} catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}	
+		} else {
 			throw new ResourceNotFoundException(id);
-		} catch (DataIntegrityViolationException e) {
-			throw new DatabaseException(e.getMessage());
 		}
 	}
